Move `operator_actor.hpp` to `exec/operator.hpp`
