+++ libtenzir/src/detail/add_message_types.cpp
@@ -19,13 +19,13 @@
 #include "tenzir/connect_request.hpp"
 #include "tenzir/context.hpp"
 #include "tenzir/detail/stable_map.hpp"
+#include "tenzir/exec/operator.hpp"
 #include "tenzir/expression.hpp"
 #include "tenzir/http_api.hpp"
 #include "tenzir/index.hpp"
 #include "tenzir/ip.hpp"
 #include "tenzir/module.hpp"
 #include "tenzir/operator.hpp"
-#include "tenzir/operator_actor.hpp"
 #include "tenzir/package.hpp"
 #include "tenzir/partition_synopsis.hpp"
 #include "tenzir/pattern.hpp"
